+++ apps/cms/apps/content/models.py
@@ -350,7 +350,7 @@ class Article(LocalizedContentMixin, LifecycleMixin):
     body = RichTextField(features=ARTICLE_RICHTEXT_FEATURES, blank=True)
     excerpt = models.TextField(blank=True)
     featured_image = models.ForeignKey(
-        "wagtailimages.Image",
+        "media.Media",
         null=True,
         blank=True,
         on_delete=models.SET_NULL,
@@ -794,7 +794,7 @@ class ProjectDiagram(models.Model):
     alt_text = models.TextField(blank=True)
     long_description = models.TextField(blank=True)
     diagram_image = models.ForeignKey(
-        "wagtailimages.Image",
+        "media.Media",
         null=True,
         blank=True,
         on_delete=models.SET_NULL,
@@ -835,7 +835,7 @@ class ProjectScreenshot(models.Model):
     alt_text = models.TextField(blank=True)
     external_url = models.URLField(blank=True)
     screenshot_image = models.ForeignKey(
-        "wagtailimages.Image",
+        "media.Media",
         null=True,
         blank=True,
         on_delete=models.SET_NULL,